+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `sh_summon` command - Summons entities with a given difficulty and blight status, along with everything the vanilla summon command offers.
+### Fixed
+- Maybe reduce the chance of capability-related crashes
+
 ## [2.0.1] - 2019-03-02
 Disabled recipe and model generators... oops. You can delete the "output" folder in your ".minecraft" folder. Sorry about that.
 ### Added